requirements.txt
asttokens==2.0.8
backcall==0.2.0
brotlipy==0.7.0
cachelib==0.9.0
certifi==2021.10.8
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.12.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///C:/ci/conda-package-handling_1649105961774/work
cryptography @ file:///C:/ci/cryptography_1639472366776/work
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
Flask==2.2.2
Flask-Session==0.4.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==5.0.0
ipykernel==6.16.0
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.3.5
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
menuinst @ file:///C:/ci/menuinst_1631733428175/work
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.13.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.8.2
python-dotenv==0.21.0
pywin32==304
pyzmq==24.0.1
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016967756/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data==0.5.1
tornado==6.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
traitlets==5.4.0
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
wcwidth==0.2.5
Werkzeug==2.2.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
zipp==3.9.0
